+++ tests-integration/test_init.py
@@ -1,9 +1,0 @@
-# pylint: disable=missing-module-docstring,missing-class-docstring,missing-function-docstring,duplicate-code,too-many-locals
-import unittest
-from certilizer import run
-
-
-class TestInit(unittest.TestCase):
-
-    def test_run_with_config_file(self):
-        run("tests-integration/fixtures/certilizer.yaml", "text", None, 20, 180)